Avoid expensive String.format for non-double values.
For group-by responses with > 20k groups it seems upto 25% of the query
latency is just spent formatting doubles while generating response.
This can be avoided in cases where we know that the result is a whole
number (especially when column data type was short/int/long).

A cleaner change would be for servers to send response in appropriate
data type, and broker to upcast to the most appropriat data type. But
that needs quite a bit of changes. This PR is a band-aid until we have a
clean fix.