On-Heap implementation of immutable StringDictionary.
The call to StringDictionary.get()/indexOf() is expensive because it has
to pad/unpad the string, create new String from byte[], perform binary
search etc. This is done in the order of numDocsScanned, per query.
It can take > 30% of server side latency if numDocsScanned is of the
order of several tens of thousands or more (99th percentile).

For String columns with very low cardinality, we can simply build an
on-heap dictionary using array + map at the time of loading segment.
This gets rid of the padding/unpadding, binary search, as well as
eliminates all temporary String creation.

Added field in IndexingConfig to specify (String) columns for which
on-heap dictionary should be loaded.

Added unit tests for the new code.