Suppress unnecessary log for partition aware routing builder.
When the retention manager deletes the segment and all servers are marked as offline
for a specific segment, the partition aware routing builder emits unnecessary logs
indicating that no server is found for the segment. However, this log is for the case
when the segment is valid while all servers are not available. Adding check in the
external view change update fixes this problem.