[TE] rootcause - related entities tracking (#1725)
* track an entity's related entities throughout the RCA framework

Whenever a pipeline emits an entity it may tag the entity with related entities it was derived from. For example, the AnomalyPipeline can now tag every AnomalyEntity with the time range, metric, and metric dimension it was found in. This solves a whole list of issues, including (a) distinction of baseline and anomaly-period events, (b) intelligent merging of multiple pipeline scores for the same entity, (c) providing the user with country information for holiday events, (d) redundant querying of data sources for context-dependent formatting, and (e) debugging of unexpected RCA results via a complete derivation tree per entity.