Improve logs during realtime segment commit in the controller and server.
Changed the controller to log the entire exception while committing the segment file.

Changed the server to log the controller host and version during split-commit upload phase,
since the segment may go to any controller host.

Re-factored the http post during split commit so that the code path is different from that
of regular get/post to the leader controller.