Compatibility patch for older http clients, a bug fix in segment state toggle, logging (#1774)
* Compatibility patch for older http clients, a bug fix in segment state toggle, logging

Some http clients do not send a leading slash in the URI in their http request.
As in "GET tables HTTP/1.1" instead of "GET /tables HTTP/1.1". Jersey framework does not
parse this correctly, so added logic to rewrite the request in such cases. Previous version
of third-eye library will now work correctly with the jersey server.

The log message, when such an API comes in, shows up as follows:
2017/08/09 10:42:14.643 WARN [HeaderAdder] [grizzly-http-server-15] [pinot-controller] [] Rewriting new Request URI http://localhost:8862/tables (incomingBaseUri = http://localhost:8862/, incomingReqUri = http://localhost:8862tables)

Also, fixed a but in the API that changes state of a segment. restlet API (and therefore
the derived jersey implementation) threw exception when changing state of all segments of
a table. Fixed in the jersey part.

Fixed a bug in the jersey implementation where toggling the state of one segment would end up
toggling the state of all segments

Also added a log of each request and response code.

* Added content-type header and src ip address to log