Cleanup: Remove unnecessary String.split() call when getting (#1785)
partition function from factory.

In prior implementation, arguments of partition function where
encoded into the string name, and there was a split on the string
to extract the name of the function. At some point we changed this
to have separate metadata for function name and argument, but missed
to remove the now redundant split on string. Cleaning it up here.