add frequency as query parameter for endpoint that update anomaly detection function, make time series retrieve endpoints do not roll-up
Anomaly detection can work on 'dimension exploration' - for a given dimension, run parallel anomaly detection for all categories in this dimension.
for example if we set anomaly detection for 'page views' and explore on 'country' dimension, and there are ten distinct countries, we are going to do
10 parallel detection, treat each country as a independent metric.

If the rollup flag is 'true', we will group those low contribution coutries (<1%) to a single category 'OTHER'. For example assume most of pageviews come from
US, CA and CN, all 7 rest have only contribute 2% together, 0.7% each. Then all the rest will be group to 'OTHER' and effectively we are doing anomaly detection for
'US', 'CA', 'CN' and 'OTHER'

Here we set the default behavior to rollup = flase, to do all possible categories in a perticular dimension.

For API calls we make it a query variable.