Plumbing for realtime consuming segment statistics (#1804)
* Plumbing for realtime consuming segment statistics

In order to allocate memory efficiently for dictionary for a column, we can attempt to use
the statistics gathered from previously completed segments.

Added a class to keep track of statistics for N previously completed segments for a table on local disk
of each server. For now, we have code to store average size (in case of a string column) and cardinality
(for all columns). These can be used when initializing a dictionary for the next segment.

Some methods are synchronized since we expect that segments may be completing at the same time for all
partitions hosted in a machine.

Added a test class.

Added methods in off-heap dictionary classes to get the statistics we want

* Addressed review comments