Wire-in config and segment stats history for off-heap memory allocation for consuming segments
Added a server config (pinot.server.instance.realtime.alloc.offheap to be set to "true") for off-heap
allocation of memory for dictionary. Currently, off-heap dictionary is always memory-mapped.

Mmap files for dictionary are in <dataDir>/<tableName>/consumers/<segmentName>.{0, 1, ...}. Multiple columns
use the same file (0.5g in size), until the entire file is mapped, and then a new file is created if any column
needs more memory.

Past history of column statistics are stored in <dataDir>/<tableName>/consumers/stats.ser. History is saved
in a circular buffer that has room for 16 segments. The file is persisted across restarts of the same server.
The handle for the stats file is created when the TableDataManager is created for that
table (i.e when the first segment for that table is assigned to a server).

The same stats object is kept around in memory until the server shuts down.
The file is modified by adding new segment statistics each time a consuming segment is destroyed.