Preparation for OffHeap config: Re-organizing parameters to RealtimeSegmentImpl
No new functionality added.

The class RealtimeSegmentImpl needs to have access to the configuration so it can determine
whether to allocate memory on-heap or off-heap. Until we turn the default to be off-heap
allocation, this will be useful to configur on a per server or per use case basis.

Re-factored the constructor for RealtimeSegmentImpl so that it has fewer arguments,
preferring to pass in entire config objects as opposed to individual elements from the
config.

Turned out that InstanceDataManagerConfig can be null only for tests, so changed the tests to include a
mock, so that this paramter can always be non-null when constructing IndexLoadingConfig.

The only behavior change in this check-in is that if IndexLoadingConfig is constructed with InstanceDataManagerConfig as null
(never happens in non-test use cases), then we will now turn _enableDefaultColumns to false because the mock for these
tests sets it to the default value.

In live situations, InstanceDataManagerConfig is never null, so the member variable _enableDefaultColumns is always taken
from whatever is configured in InstanceDataManagerConfig