Enable Jersey API as the default, restlet as backup (#1813)
* Enable Jersey API as the default, restlet as backup

* Modified UploadRefreshDeleteIntegrationTest for more predictability

The testRetry() method was building a segment, computing its length, removing the
segment and invoking buildSegment in another thread and trying to push the newly
built segment.

The problem with this approach is that the length of the file the second time around can
differ from the length of the file the first time. And we keep retrying with the wrong length,
and it fails all the time.

Since the objective of the test is to test that the retry logic works, knowing that it will fail
the first time, it is better to rename the generated segment to a temp file, and rename it back
after some sleep time.

This way, the first iteration will not find the file, and second (or maybe third) may end up finding
the file to send.