Avoid unnecessary split/join for in-clauses.
Current broker code joins the values in in-clause with '\t\t', that
requires server to get the individual values using String.split(). This
becomes extrememly inefficient for queries with large number of values
in the in clause (hundreds). This change elimiates this unwarranted
join/split, thus improving perfomance of affected queries. To avoid
breaking any broker/server compatibility:

1. Added config (off-by-default) in broker to not join the values, and
send them as list. Server checks if the size of list > 1, then
recognizes then skips split, or else falls back to original behavior.

2. Brokers and servers can be deployed with this feature off. Once all
are updated, the feature can be turned ON in broker, and ther servers
would already have been updated to handle the new protocol.

3. The config will be removed eventually to keep the code clean.

Added unit test for the new code.