Fix NPE and add a test case for older history files
If we fail to read the stats history file for any reason, we would end up getting an NPE because _statsHistory
went uninitialized. Fixed to create an instance with a new file.

Also added a v1 serialized file to the repository so that we can ensure that future changes to the class
can still read v1 files.