Fix time granularity in anomalies resource (#1829)
The getTimelinesViewInMonitoringWindow() takes only the bucket time granularity from dataset configs. The function also has bucket time granularity from user. This design fails when the two config are not the same. This pull request is to resolve this issue. It takes both definitions of buckets, and apply the definition from function in default. But, it also compares the definition from dataset. If the function's is smaller than the dataset's, it applies the dataset's instead.

Another part is to allow users to update function bucket time granularity via the function update endpoint.

* Resolve the time granularity issue when generating function baseline

- Original design takes the time granularity info from dataset config, but it is not reasonable.

- New design takes the time granularity from function, but it also compare the time granularity from dataset config. If the time granularity of function is smaller than the dataset's, use the dataset's instead.

* Update anomaly function update endpoint

- allow user to update function bucket size and unit