Retry segment download only on 5xx error codes.
When a segment download fails with 4xx (esp 404) no amount of retry is going to help us get
the segment. We waste significant time (30s per segment) if we reach such a condition, unable to
report startup complete because the state transition is still being attempted.

Not to mention that a thread is held up sleeping for a long time.

Introduced a new PermanentDownloadException that callers can examine before attempting a retry.