Clean up memory map files for a segment on start
If a host restarts, we will end up throwing an exception by trying to create files
with the same name for the mmap.

Added code to remove the files upon start (i.e. creation of new instance of mem manager for a segment),
and added a test for it.

Also added segment name to the logs so we can get an idea of how much mem is being allocated