Make number of segments that can be refreshed simultaneously
We introduced a lock on the helix threads refreshing a loaded segment, so that we do not run out of memory
with too many segments loaded. However, if segments are loaded via MMAP and the server is in a multi-tenant
(low qps) cluster, it should be fine to overload via MMAP for a brief period.

This commit makes the number of simultaneous threads configurable. A value of 0 (or less) indicates unlimited
number. Any value above 0 configures the maximum allowed threads to refresh.

Default is 1, preserving the old behvaior of a lock.

NOTE: While setting this number to a higher value (or 0) may increase segment load time, there will be more
      threads occupied loading segments, so query latencies may actually increase during refresh time