[PINOT-5281] Implement relabance API on the controller (#1928)
* [PINOT-5281] Implement relabance API on the controller

Added methods in the controller to rebalance realtime and offline tables.

For offline tables, we rebalance all segments amongst the new instances provided.

For realtime tables, we only rebalance ONLINE LLC segments. The CONSUMING LLC segments
are re-balanced when ValidationManager detects new number of replicas. If all replicas are
OFFLINE then the entries are there only for debugging, so we can leave them as is.

Handled changing number of replicas as well (either increase or decrease).

The numReplicas field in the Idealstate is updated after the rebalance operation,
so that manual update of zk Idealstate node is avoided.

ALERT: Changing the number of replicas AND number of instances with a single rebalance is NOT supported.

Added tests

* Addressed review comments