[PINOT-5350] Protection against file system timeout when fetching segments (#1933)
It is possible that the controller times out while reading from a (possibly remote) file system.
In this case, the controller already sends out a 200 OK, but the request times out, closing the socket
connection to the server. The server assumes the file received is complete, but untar fails, and an exception
is thrown, resulting in the segment transitioning to ERROR state.

While we can add retries on specific exceptions from untar, it is best to send the content length from the
controller and compare it with the length of the received file in the server. The existing retry mechanisms
will then kick-in, and the file may be received correctly on a subsequent retry.

Tested with the controller not sending the header as well, in case there are servers that talk to the older
controllers