Ignore CONSUMING state in externalview while reporting segment status
It is possible for Idealstate to be in ONLINE and externalview to be in CONSUMING
state (for realtime LLC segments) for a short period of time. The segment status
checker runs every 5 minutes, and for tables that produce segments fairly frequently
it is possible that at any given time, some segment is in this combination of states.

Ignore the CONSUMING state in externalview so that we report alerts correctly.

If a segment is in this combination of states for too long, the alerts from the
server should fire, regarding non-consuming kafka partitions