Fix for a segment having no statistics
If a table is dropped when no rows have been consumed, the statistics will get filled with zeros
and that can cause an index-out-of bounds exception the next time table is created and
consumption starts.

This is because the RealtimeTableDataManager is not removed when a table is dropped.

There are also other conditions that can cause 0 stats to happen, so this PR ensures that
stats returned is never 0.