Add support for virtual columns
Add support for virtual columns, which are columns that have no on-disk
representation but instead contain values that are filled in at run
time.

Add support for three built-in virtual columns:
 - $docId: Returns the document id for the current row
 - $segmentName: Returns the current segment name
 - $hostName: Returns the hostname of the current server

These columns can be used for debugging purposes, such as debugging
query routing (which server answered which query) and data issues (what
are the data ranges in each segment).

Additional built-in virtual columns can be added to
VirtualColumnProviderFactory and customized virtual columns can be added
to any schema by adding a provider to the classpath and specifying the
fully qualified class name as the "virtualColumnProvider" property of
the field specification in the Pinot table schema.