Fix off heap inverted index creation on MacOS X (#2015)
Fix the off heap inverted index creation on MacOS X. It seems that on
MacOS X, the behavior for the position on a file channel is to be
absolute while on Linux it is relative to the current file position.

Rewrite the transferFrom call to use an absolute position on all
platforms.