Split overloaded api 'indexOf' in dictionary into 'indexOf' and (#2022)
'insertionIndexOf'.

The existing indexOf api on Dictionary interface performs binary search
to find out the actual index of the object, or possible insertion index
if the object does not exist. Calls to indexOf that are not interested
in possible insertionIndex of the object can be optimized by making the
dictionary map based, and thus avoiding binary search.

Renamed existing 'indexOf' to 'insertionIndexOf' that preserves existing
behavior. Added new 'indexOf' that returns -1 if object is not found.

Enhanced OnHeapStringDictioanry to use a map for 'indexOf' and thus
avoid unnecessary binary search.

For rest of the implementations, 'indexOf' currently just calls
'insertionIndexOf', and overwrites the result to -1 if object not found.
There's no additional penalty of doing so. But this opens the door for
other implementatinos to also make use of map based lookup in future.