Add request ID to broker connection logs. (#2027)
* Add more descriptive names to the future handles, for logs.

We see a possible memory leak in the broker. The likely flow when this happens is as follows:
- The broker attempts to send the request to (say) 5 servers (in 5 different threads)
- It cannot get a connection for one of the servers, so it marks the future as Error future and returns to the caller.
- When the future is marked as an error, all the other futures are cancelled. Not sure which thread this happens in
- By the time they are cancelled, it is possible that some of the servers have already responded.
- The broker ignores these responses since the request has been cancelled, and a leak results.

Over time, this can manifest as netty erroring out while trying to get direct memory.
Not sure if this can also cause netty to run out of threads, but that could be a seconday effect because
we run out of threads trying to reclaim memory? (we have seen an increase in number of threads when this
happens).

The fix seems to be to release netty's ByteBuf when cancel is called after the response is received
(or, when a response is received after the cancel is called). However, it is not clear at the moment whether
there is another thread that may race to access the response's byte buffer after it has been released.
So, the code is commented out for now.

The additional logs may help us identify the problem better next time.

* Added the context while creating AsyncResponseFuture in KeyedPoolImpl