Limit the number of segment builds on a server
Default behavior is to keep this unlimited. However, in some use cases that may have
a loq latency requirement, with a high QPS, and also a large number of columns, we
may incur heavy GC pauses during segment builds.

The only down-side of enabling this limit is that we will pause the consumption if
segment build is delayed on the committer. Since we are consuming on a per-partition
basis, the consumption should catch up quickly.

Alternative is to add more servers.