Added an option to allocate Direct Offheap memory for realtime consuming segments (#2038)
* Added an option to allocate Direct Offheap memory for realtime consuming segments

So far, off-heap allocation for realtime consuming segments meant MMAP-based allocation.
This may generate excessive disk i/o depending on the configuaration. Added a setting
to allocate this memory via direct allocation.

Since direct memory translates to an internal memory allocation call, we may end up
allocating more than what we actually need. Fragmentation is also possible, since multiple
threads will be allocating memory of different sizes, and may release them in different order.

A more fancy algorithm may be to do a bit of memory management in this code (like how mmap
does, but at a smaller mem level, perhaps). For now, we leave it at this.

Changed some logs to be more informative, and added one metric on the server to have it
report the number of segments scanned in a query as a gauge.

Tested LLRealtimeClusterIntegrationTest with both direct and mmap allocations

* Added a clarfication comment