Handle consumder dir not configured, and cleanup old consuming mmap files if any
If the consumer directory is not configured, when we create the consumer dir in some
unknown place (depending on the path env of the application). Changed it to be under the datadir.

Also, it is possible that segments may be re-assigned after a server restart, causing old
mmaped segments to remain (if the last shutdown was a hard kill of the process).
It is best to clean out all mmap files under the consumer directory when the table data manager
is inited.