[TE] rootcause - metric analysis pipeline with quantile predictor (#2036)
The existing MetricCorrelationRankingPipeline makes a number of assumptions about the type of metrics being compared that no longer hold in practice. This PR add a more generic MetricAnalysisPipeline that allows related metric to be ranked with more generic strategies. The PR furthermore adds small missing features and fixes bugs that turned up during the implementation of this pipeline.

* adds MetricAnalysisPipeline

* adds quantile-based ranking strategy

* adds quantile function to dataframe double series

* adds helper to fetch timeseries with aligned timestamps

* fixes race condition in thirdeye request builder