Enable support for splitting predicate values for NOT-IN clause. (#2188)
Support for splitting predicate values (not joining using '\t\t') is
only enabled for IN clause. And to make this feature always on, we need
to first enable it for NOT-IN clause as well, so server side can be
backward compatible and recognize both splitted and joined predicate
values. Once that is done, we can enable the feature by default.

1. Added support in NOT-IN predicate to handle both joined as well as
splitted predicate values.

2. Extracted out a base class to move common code for IN and NOT-IN.
Renamed 'getInRange' to getValues() so it can be moved to base class as
well.

3. Updated test to cover the NOT-IN case.

Once server side is available to handle both cases, will file another PR
to enable always splitting the values.