[PINOT-5458] Remove static initialization of classes in SegmentFetcherFactory (#2238)
* [PINOT-5458] Remove static initialization of classes in SegmentFetcherFactory

Static initialization of classes in SegmentFetcher factory can have unneeded exception stack trace
when the system starts up. The HttpsSegmentFetcher, for example, will throw (and log) initialization
exceptions, even if it is overridden with some other class. This poses a problem for deployments with
automated exception detection, not to mention unnecessary exception logs confusing the admin when
they have configured a different class.

Removed static initialization in the SegmentFetcherFactory class.

If initSegmentFetcherFactory is not called, then map of protocol to SegmentFetcher is not populated.

* Addressed review comments