Added a method to get a status description from ServiceStatus (#2315)
Often happens that a service is not finished starting up because of a specific table or
partition being unavailable as yet. Since we don't log each time we return a STARTING status,
it is good to keep track of what we are waiting for and provide an interface to retrieve it.