Minimize window in which two controllers can update metadata for LLC segment commit (#2321)
* Minimize window in which two controllers can update metadata for LLC segment commit

It is possible that the leader controller undergoes heavy GC after accepging the
segment commit request from a server. In this case, the leader could lose zk connection
for a while and wait to regain connection in order to test leadership.  During this time
another controller could gain leadership and commit the metadata, with a slightly different
kafka offset, and create a new segment. Meanwhile, the former leader, without knowing that it
has lost leadership, could go ahead and commit another new realtime segment, which could
have a different name than the segment committed by the new leader!

We minimize the window of this happening by checking one more time right before we commit
metadata.

Also, we check for leadership as well as zk connectivity instead of just leadership check.

Added unit tests

* Addressed review comment