[TE] rootcause - extended metric analysis (#2279)
This PR extends the RCA backend to support MetricEntities with a filter tail in the URN. Adds a specialized pipeline for mapping between metric entities that also translates filters if possible. Furthermore extends the one-step dimension analysis pipeline into an iterative component/factor analysis. Also adds a complement of unit tests and helpers.

* long-format metric urns ("thirdeye:metric:12345:filter=value:otherFilter=otherValue")

* support long-format urns in metric-related RCA pipelines

* iterative metric component analysis pipeline (restricted PCA)

* exclusion filters in thirdeye-requests ("filter != value")