Added on-heap dictionary support for int/long/float/double data-types.
For low cardinality columns, it might be more efficient to store the
dictionary directly on-heap, as long as the heap foot-print does not
increase. The advantages are two fold:

1. Provides constant time lookup, as opposed to binary-search.
2. Saves on redundant ser-de of same values from byte-array to
destination data-type over and over again.

Heap usage increases long with cardinality of the item, so it is not
advisable to use this feature where cardinality is medium/large, or can
grow, unless the heap usage is acceptable.