[TE] rootcause poc - support special characters in urns (#2345)
People are horrible, so we need support for special characters in urns (with filter tail). Depending on the dataset, filters may take values that are misinterpreted as URN tokens, e.g. "pageId=my:app:/route/subroute?q=text". This PR adds frontend and backend support for URI-encoding URN fragments so that urns can be tokenized safely.