Re-factor PinotLLCRealtimeSegmentManager methods to reduce the number of arguments. (#2404)
* Re-factor PinotLLCRealtimeSegmentManager methods to reduce the number of arguments.

The argument count seems to increase each time we want something new. Passing around
tableConfig and kafkaStreamMetadata seems to obviate the need for most other arguments.
In another step, maybe we can also get rid of KafkaStreamMetadata, since that should also
be (technically) got from tableconfig.

There is still a dependency with the kafka consumer, so we still need to be careful in tests
ensuring that we don't start to initiate a connect to Kafka. At some point, we should
pass in a mock kafka consumer, so that we can can get the offsets we want from the
consumer. We are only interested in offsets for these tests.

* Addressed review comment