Support updatable metrics during realtime consumption.
One of the reasons that serving latency of realtime nodes may be worse
than the offline counterparts is that data can (and almost always is)
aggregated in the offline pipeline, reducing the number of rows to be
processed.

For cases with high Kafka event rate and with duplicate dimension
values, this performance delta can be huge, impacting the overall
latency.

This PR addresses this problem by allowing for updating of metrics for
rows with same dimension values. The implementation is as follows:

1. We maintain an off-heap HashMap with FixedIntArray (containing
dictIds of dimension columns) as key, and docId as the value.

2. When a new row comes in, we check if the dimension values are present
in the map:
   If yes, then we fetch the docId from the map, and update the metric
columns.
   - If not, then we index the new row.
   This features requires metric columns to use no-dictionary indexing.
The reason being updating of metrics may leave unsused values in
dictionary that need to be cleaned up before segment flush. Using
no-dictionary indexing eliminates the problem completely.
   Also, this feature is only supported for LLC consumer.

3. The feature is off-by-default and can be enabled with a config
   - TableConfig.IndexingConfg.updateMetrics

4. Added unit tests for the new feature.

5. Performance gains measured by this feature using real production
data:
   - An average of 4x reduction in numRows. Although, more common
     dimension combinations (that are also frequently queried) see upto
1000x reduction in number of rows.
   - Realtime nodes now perform almost at par with offline nodes for
     query serving.