Cleanup: Enable splitting of in-clause strings by default.
PR's #1819 and #2188 enabled optimization that avoids unnecessary
split/join of strings in IN/NOT-IN clauses. This was done in a backward
compatible way where server side could handle both split as well as
unsplit strings.

This optimization was off-by default and was enabled with a config.
Now that the changes have been around for several months, this PR cleans
up the config and enables the change by default.

The server-side is still backward compatible with old broker. This can
be cleaned up future PR once broker changes have been around for a
while.