[TE] rootcause - anomaly baseline support (#2648)
After multiple iterations on the anomaly baseline endpoint this PoC provides support for embedding the computed anomaly function baseline in the chart. Support is orthogonal to the existing user-selected baselines and can be toggled by the user in the legend. The current implementation hosts the logic entirely in the RCA controller and replaces the default metric baseline time series when enabled.

* add toggle for displaying detection baseline rather than default baseline

* service for retrieving data from anomaly function baseline endpoint

* rewrite route for anomaly entry point

* support anomaly function urns in frontend utils

* support highlight of multiple time series at once

* compensate in backend for superfluous/missing pre-aggregation keywords in stored anomalies