Included more parameters in realtime server http requests to the controller.
In preparation for auto-tuning of consuming segment allocation in multi-tenant
realtime cluser:

The realtime segment completion requests posted to the controller now include
the following parameters as appropriate:
        memory used
        number of rows consumed
        build time
        wait time to acquire the build semaphore.

Broadened the scope of SegmentFileAndOffset class (renamed it to SegmentBuildDescriptor) to
include more than just the file name and offset. It now serves as a return value of
buildSegmentInternal() method, and has the build time and wait time as well.