Addressed comments from design review for segment size based row threshold updater (and more).
Review comment: Added logic to set a minimum of 10 k rows if the target number of rows computes to be too small.
This avoids corner cases where incoming stream has issues and we consume too small number of rows.

Changed updateFlushThreshold method to be synchronized since multiple partitions could be calling invoking the same method
in the same instance of the object, and we need to serialize them in order for the ratio to be correct.

Changed the threshold updater to accept input from partitions other than patition 0 if there the current
ratio is 0. This helps us jump start all consuming partitions sooner during table start up, or controller
switch of mastership.

Added logic to start a fresh threshold updater when a table is created. If a table is removed and recreated
(e.g. during experimentation), we get the old threshold values (not desirable)

Updated log messages to contain the correct value of target num rows.

Updated log messages to contain the new segment name for which target num rows is being computed.

Added test cases