[PINOT-6628] Add instance ID when reporting stopped consumption (#2940)
Include instance ID when placing an http call to the controller after encountering permanent stream
consumption errors.

If the instance ID is not included, the controller ends up changing the Idealstate to include
"UNKNOWN_INSTANCE" in OFFLINE state. If at least one instance continues to consume, then this is
not a fatal problem, because eventually that instance will complete the segment and all instances will
go to ONLINE state.

However, if all instances face stream consumption issue, we will have an issue where consumption has stopped
but the state is still CONSUMING. This will trigger consumption alerts, but the auto-correction logic will
not be able to detect the situation.