Lower the probability of intermittent test failure for LargeClusterRoutingTableBuilderTest
This test fails occasionally when the variance between least loaded and highest loaded server
is more than 50% across all routing tables. Since we have a random generation of routing tables
it is not guaranteed that this is always the case.

Lowered the probability of failure further by testing the criteria 5 consecutive times before
asserting a failure.