[TE] frontend - aaronucsd/Added new shared summary email feature
Summary:
Allow for a share feature from the dashboard. THe user can export the shared summary as pdf,
copy to clipboard and share a snapshot link. There is a way to hide metric and alert level from the final shared content.
The shared feature allows for high level comment and metric level comments.

1. fixed Wow to be predicted instead on dashbaord and shared summary
2. Added new ember data for shared summary
3. added new libs to support PDF download
4. added cryptojs for hash support