Reducing garbage when indexing multi-value columns (#3048)
The dictionary ID of multi-valued columns need to be in sorted order for any given docId in the fwd index.
In the dictionary creator, when adding the multi-value field, we already create a new array. It makes
sense to sort the values right there, and assume that the values coming into the fwd index builder
are always sorted. This way, we avoid creating another array for multi-valued columns during
segment creation.

Verified that there are no other callers to te fwd index builder, and also updated javadoc to reflect
the new expectation of the input.