Reduce heap allocation during realtime segment creation (#3057)
In order to calculate min and max values of a no-dictionary column we walk through the entire value set,
allocating a good bit of heap memory. But then we really don't use the value generated.
The getMin() and getMax() methods are called only to construct ColumnIndexCreationInfo (the values
gathered from columnStatistics are stored in the ColumnIndexCreationInfo object). But then
the min/max values from ColumnIndexCreationInfo are retrieved only for the time column and not for
any other column. Time column is never a no-dictionary column, so we can remove all the code in the
no-dictionary column to compute min-max.

This essentially reverts #1468, which was submitted because of review comments on #1462.
Comment on #1462 should now be addressed, because we now throw exception instead of returning
some invalid value.