Refactoring Pinot segment upload (#3153)
* Refactoring the upload code to be a set of interfaces that have different implementations across different upload types and universal utils classes that are used for all types.

* Makes it easier to add new features, such as ways to encrypt/decrypt segments.

* Added an integration test for URI upload. Note that this upload method intentionally behaves differently from the old URI upload in that it always will move the segment to a controller-dictated final directory.

* URI upload will always go through the JSON content-part endpoint since the multipart endpoint will throw errors

* Changed the current multipart endpoint and json segment upload endpoints to go through my new upload method.

Future work

Add metadata upload type as an optimization
In this PR, I removed support for segment names with strange characters. The reason for this is because we are now using a URI abstraction. That means that the final segment resting place will have to fit into the URI abstraction and will require the filename to be encoded as well. Will need to do a bigger change to support that, aka, store segments as their encoded names both in zk metadata and in their file paths/look into that more and did not want to do that as part of this pr.