Convert byte[] defaultNull values into String equivalents during (#3196)
Schema serialization.

Since defaultNullValue is stored as an Object in FieldSpec, and for
BYTES data type the Object type is byte[], we need to convert it into
String value when need. This also helps serialize and deserialize the
schema correctly.

Added unit test to cover serialization/deserialization of schema
for various defaultNullValues.