Disable LLC validation when LLC is not enabled (#3381)
We attempt to validate llc segments of a table even if the table has only hlc segments.
Part of the exercise of validating llc segments is to check if number of stream partitions have
increased. For Kafka llc, this check needs a broker config (Kafka hlc does not need
broker config). Since the broker config is missing from stream config, we end up throwing an exception
in the controller.

This creates an unnecessary exception log for these tables.

The exception is also visible if we let quick-start-realtime.sh run long enough, creating a bad user
experience.

There are no other side-effects to this exception. We log the exception and move on to validate the next
table.