Add LimitStringLengthRecordReader to limit the length of string columns
Motivation:
Recently, we run into issues where bad data is getting indexed into Pinot.
The length of one single string value is more than 1M, and cause the
dictionary size to be huge (>2G) and also cause int overflow and then crash
the server. In most of the cases, super long string is not even queriable,
and trying to index them is usually caused by client code bug. So we decide
to add a configurable limit to the length of string columns. If the value
for the column goes over the limit, the value will be trimmed to be the
length limit.

1. Added "stringLengthLimit" into FieldSpec, so each field (column) can have
  different length limit
2. The default limit is 512 (string length but not decoded bytes length).
  The feature can be disabled by setting the limit to a value <= 0
3. Added LimitStringLengthRecordReader that works on the existing record
  readers but trim strings that are longer than the length limit