Add max length limit to string values
Motivation:
Recently, we run into issues where bad data is getting indexed into Pinot.
The length of one single string value is more than 1M, and cause the
dictionary size to be huge (>2G) and also cause int overflow and then crash
the server. In most of the cases, super long string is not even queriable,
and trying to index them is usually caused by client code bug. So we decide
to add a configurable limit to the length of string columns. If the value
for the column goes over the limit, the value will be trimmed to be the
length limit.

Added "maxLength" into FieldSpec, so each field (column) can have different
length limit.
The default limit is 512 (number of characters but not decoded bytes length).

NOTE: for string with null character, the part after the first null character
will be truncated because null character is reserved as padding character