Handle invalid durations in the duration DSL
Handle invalid durations in the duration DSL by removing invalid values
when loading (eg. "null DAYS" or "null null"). These values come from
invalid old-style configurations that got migrated to the new config
format.

This does not fully round-trip. For example, an old config with timeUnit
of DAYS and timeValue of null will lose the DAYS qualifier, although
this configuration was invalid (and ignored) in the first place.