Added a log message on pinot-server when controller leadership change is read from zookeeper. (#3522)
Recently we saw a case where the server was repeatedly reporting that it received NOT_LEADER
responses from the controller. On further look at the logs, it turned out that the controller
did not get a helix notification that it became leader, so the server was logging what it saw
in zookeeper.

The behavior did lead us into looking at other things before we found the root cause.

This log will clearly identify how the server concludes who it thinks the leader is.