[TE] frontend - harleyjj/create-alert - YAML autocomplete for create … (#3617)
The autocomplete works for metrics only until a metric has been chosen. Then it replaces the whole text field in the editor and enables autocompletion for dimensions and filters (dimensions and filters will only be offered for the autocompleted metric, and they will only replace the string instead of the whole text field).

It might be better to merge it after we have the componentized YAML editort, but I just thought I'd put the PR up for code review.